common -- refactor axis_index to a single file
