+++ rstsr-common/src/lib.rs
@@ -6,6 +6,7 @@
 pub mod prelude;
 pub mod prelude_dev;
 
+pub mod axis_index;
 pub mod error;
 pub mod flags;
 pub mod format_layout;
